+++ src/containers/Shop.tsx
@@ -13,7 +13,8 @@ import groupBy from "util/groupBy";
 import ShopList, { ShopItemTuple, ShopItemType } from "views/shop/ShopList";
 import { BaseItemProp, ItemProps } from "views/shop/types";
 import { Maybe } from "graphql/jsutils/Maybe";
-import { BfsId, Quantity } from "global/types/types";
+import { Quantity } from "global/types/types";
+import { BfsId } from "global/types/identity";
 import { ripLoadObject } from "util/ripLoadObject";
 import windowStore from "../data/WindowStore";
 import partition from "../util/partition";
