+++ src/data/useFriendLO.ts
@@ -1,6 +1,6 @@
 import FriendStore from "./FriendStore";
 import useFluxStore from "./useFluxStore";
-import { BfsId } from "../global/types/types";
+import { BfsId } from "global/types/identity";
 
 function useFriendLO(id: BfsId) {
     return useFluxStore(() => FriendStore.getFriendLO(id), [FriendStore], [id]);
