+++ src/features/RecipeDisplay/hooks/useLoadedPlan.ts
@@ -3,7 +3,7 @@ import planStore from "features/Planner/data/planStore";
 import React from "react";
 import Dispatcher from "../../../data/dispatcher";
 import PlanActions from "features/Planner/data/PlanActions";
-import { BfsId } from "../../../global/types/types";
+import { BfsId } from "global/types/identity";
 
 export const useLoadedPlan = (pid: BfsId | undefined) => {
     // ensure it's loaded
