--- src/global/types/recipe.ts
@@ -1,3 +1,0 @@
-import { Recipe } from "__generated__/graphql";
-
-export type DraftRecipe = Omit<Recipe, "owner" | "subrecipes" | "favorite">;
