+++ src/global/types/types.ts
@@ -1,15 +1,5 @@
 import { CheckableActionType } from "util/typedAction";
-
-export type BfsId = string | number;
-
-export interface UserType {
-    id: BfsId;
-    name: string | null;
-    provider: string;
-    email: string | null;
-    imageUrl: string | null;
-    roles: string[];
-}
+import { BfsId, UserType } from "global/types/identity";
 
 export interface Ingredient {
     id: BfsId;
@@ -17,18 +7,31 @@ export interface Ingredient {
     type?: "Recipe" | "PantryItem";
 }
 
+export interface Recipe extends Ingredient {
+    calories: number | null;
+    directions: string | null;
+    externalUrl: string | null;
+    ingredients: IngredientRef[];
+    labels: string[];
+    photo: string | null;
+    photoFocus: number[] | null;
+    totalTime: number | null;
+    recipeYield: number | null;
+}
+
 export interface PantryItem extends Ingredient {
     storeOrder: number;
 }
 
 export interface IngredientRef {
+    id?: BfsId;
     raw?: string;
     quantity?: number | null;
     preparation?: string | null;
     units?: string | null;
     uomId?: BfsId;
-    ingredient?: string | Ingredient | null;
-    ingredientId?: number;
+    ingredient?: Ingredient | string | null;
+    ingredientId?: BfsId;
 
     /**
      * On the shopping list, napalm entries will only have name, not raw.
@@ -40,6 +43,34 @@ export interface Quantity {
     quantity: number;
     units?: string;
     uomId?: number;
+}
+
+export interface RecipeFromPlanItem extends Recipe {
+    subtaskIds: number[];
+    subrecipes: Recipe[];
+}
+
+export type Subrecipe = Pick<
+    Recipe,
+    "id" | "name" | "totalTime" | "ingredients" | "directions"
+>;
+
+export interface FullRecipe {
+    recipe: Recipe;
+    subrecipes: Subrecipe[];
+    mine: boolean;
+    owner: Omit<UserType, "provider" | "roles">;
+}
+
+export interface SharedRecipe extends Pick<Recipe, "id"> {
+    secret: string;
+    slug: string;
+}
+
+export interface DraftRecipe extends Omit<Recipe, "photo"> {
+    photoUrl: string | null;
+    photoUpload: File | null;
+    sourceId: string | null;
 }
 
 export interface FluxAction {
@@ -55,3 +86,5 @@ export interface Page<E> {
     last: boolean;
     content: E[];
 }
+
+export type FormValue = string | number | number[] | string[] | File | null;
