+++ src/providers/Profile.tsx
@@ -5,13 +5,16 @@ import React, {
     useCallback,
     useContext,
     useEffect,
-    useState,
+    useState
 } from "react";
 import LoadObject from "util/LoadObject";
-import { API_BASE_URL, LOCAL_STORAGE_ACCESS_TOKEN } from "../constants";
+import {
+    API_BASE_URL,
+    LOCAL_STORAGE_ACCESS_TOKEN
+} from "../constants";
 import GTag from "../GTag";
 import { Maybe } from "graphql/jsutils/Maybe";
-import { UserType } from "../global/types/types";
+import type { UserType } from "global/types/identity";
 
 // global side effect to ensure cookies are passed
 BaseAxios.defaults.withCredentials = true;
